Refactoring day 1 solution
Rather than return a Pair of ints for each line,
return a list.

Then turn the lists from rows to columns by
flatMapping the parsed lines and grouping by
the index in each line. This results in an
implementation that is still comprehensible, but
avoids manually appending to multiple mutable lists.

The slightly simpler parseLine function now fits
nicely on a single line.

Improve countOccurrences by using groupingBy and
eachCount instead of groupBy and a custom lambda
(this would also be more efficient for really large
inputs)