Day 2, part 2
For cases where the levels aren't immediately valid,
loop through the levels one at a time and construct
a list of levels without that input and try again.
If any of those work, return true.

This is O(n^2), but still performant enough, and
there are some tricky instances in my input that
would make solving the problem in a "cleverer" way
a bit tricky, such as when the two first numbers in
the sequence match.