Minor optimisation for Day 2, part 2
Realised that we don't need to run isSafe at the
beginning of isSafeWithDampening, because if all
the levels are safe, removing the first level will
definitely produce a list of levels that works.

Split the happy path tests for isSafeWithDampening
into two tests to reflect "passes as is" and "only
passes with a bad level dropped" as two separate
cases.