Day 4, part 1
Find all X characters, and then proceed in all
directions looking for an M. When you find one,
proceed in the same direction looking for an A and
then an S. Count the results.

Stymied by a really stupid bug in my Point
extension functions (grr), which is now fixed.