Day 4, part 2
Crossed MASes have an A at the center, so start by
finding those.

Then to find the cross, iterate through the diagonal
directions, looking for a direction with M that has
an opposite direction with an S. If there are
exactly two of those, it's a MAS cross.

(Crosses on the cardinal direction are MAS Tees and
don't count. I learned that the hard way.)

Implement DiagonalDirection and an opposite function
in utils.

Find all X characters, and then proceed in all
directions looking for an M. When you find one,
proceed in the same direction looking for an A and
then an S. Count the results.

Stymied by a really stupid bug in my Point
extension functions (grr), which is now fixed.