Improve Day 5
The index data structure used for the comparator in
part 2 can also be used to solve part1: given an
index that will return the set of pages that must
come after a given page, build a set of pages
visited so far, and if any of the pages that come
after are in that set, it's invalid. Add the current
page to the set of previous pages as you visit it.

Move isUpdateInRightOrder onto PageRuleComparator
to be able to access the index, and allow the index
to be easily reused across invocations.

Also move fixPageOrders onto the class for
conciseness.