Day 6, part 2
For every point in the grid that isn't already an
obstruction or the guard's starting point, try to
add an obstruction.

For each potential solution, detect a loop by
tracking the PointAndDirection the guard has
previously entered for this version of the grid, and
if the guard reaches one it has reached before,
there is a loop. If the guard leaves the grid, there
is no loop.

Solution is slow: 4.5 seconds of runtime.