Day 8, part 2
findAntinodesBetweenWithResonantHarmonics
calculates the difference between the two
points and proceeds in each direction until it
leaves the map.

Overload findAntiNodes to use either operation
and use the part2 operation for part 2.

Parse the set of antenna locations for each
frequency. Generate the two-element
permutations for each set, and then find the
points in line with both that are within the
map boundary.

Count the unique ones across all frequencies.