Day 9, Part 2
Build a list of files that haven't been
moved, (I don't think my modelling choice was
great today, as this was a little bit
difficult) and then maintain a tree of free
disk chunks to efficiently find the first
open space that's big enough.

Solution runs in just over 350 ms.