Optimise Day 9
Move to TreeMap + PriorityQueues for indexing.

The "right way" to do this in a system where
the file sizes could be considerably more
variable than 0-9 blocks would be to use a
TreeMap to find free space.

Similarly, for the earliest free chunk in
each size a PriorityQueue is the "right"
data structure to use.

However, performance is unchanged.