Day 12, Part 2
The less said about this, the better.

Introduced the concept of a SideSegment,
which has a plot (Point) and a border.
Use the points set to find the edge
segments, and then find unique edges by
going row by row and looking for segments
that differ by more than one (1 by 1 for
vertical segments).