Simplify Day 12, Part 2
The 4 stream operations used to count
unique north, south, east and west edges
were close to duplicate. Refactor them so
that they are *exactly* duplicate, and
then remove the duplication.

This requires introducing an extractor
function so that we can use
segment.plot.x when counting vertical
edges, and segment.plot.y when counting
horizontal edges.