Day 14, Part 2
After each round, use the region
exploration approach from day 12 to
find adjacent robots.

The idea is that a christmas tree
shape is a big clump of adjacent
points. Run for a bunch of rounds
(figured out 7500 is good by
gradually increasing), tracking the
round with the biggest clump. Print
it out to check.

Sure enough, the round with the
biggest clump has a Christmas tree
in it (only in production data).

Solution runs in about 1.5 seconds