Day 15, Part 2
Tested all the edge cases, fixed bugs, and
it works.

Moving the robot east and west is basically
like part1. For north and south, use a
flood fill algorithm to find all boxes that
would be pushed and make sure none of them
are obstructed by a wall.

Solution isn't super fast: runs in < 800ms.