Day 16, Part 2
Extend the utility Dijkstra's algorithm to also
track the previous node — or nodes — that have
the lowest cost for each node.

Use this information to derive the shortest paths
through the node (being sure to only use the
cheapest end state, since we're using
PointAndDirection) and return the count.

Solution runs in just over 150 ms.