Day 17, Part 2
I did not even remotely test drive this solution.

When looking at the output values as you increase
register A starting from zero, you see that they
eventually start outputting more results, and that they
occasionally generate results that match the rightmost
parts of the desired output. Turns out this is in
powers of eight — 3 bit computing appending a value
each time.

Find a solution that generates the rightmost digit.
Shift this to the left by three bits and then try 0..7
again from there looking for a solution with the two
rightmost digits matching, etc. This got up to about
5 matches until it failed... because if there are
multiple options that yield a value only one of them is
the right one.

Turn this into a recursive depth-first search function:
given a register A value that has worked so far and the
number of outputs matched so far, shift it left by 3
bits and try from 0-7: if any aRegister value matches
an additional output digit when executed, try it
recursively. If no options match this is a dead end.
Once all digits match, this is a successful input.
Given more than one successful route take the smallest
one.

Surprisingly, this solution is really fast: it
generates the solution to both parts in 16 ms.