Day 20, Part 1
Use Dijkstra's algorithm from the end to the beginning.
Then find potential cheat exits (points with a
manhattan distance of 2 from a point on the original
path from start to end), and use the costs
map from the Dijkstra's search to find exit points that
are at least 100 cheaper than the start point,
including the steps used in the cheat itself.