Day 20, Part 2
Create a part 2 count that uses the same approach as
part 1, but looks for any exit points that can reach
the end that are within a manhattan distance of 20
points from any point on the original line.

Use the dijkstra search costs to find cheats that have
the required costs savings and return the count.

Solution runs in ~1400 ms.