Day 21, Part 2
Same as part 1, but increase the number of "rounds" to 25 + 1
for part 2.

Even the new approach slows to a crawl with 25 robot keypads.
Add caching for intermediate results, which makes the
solution lightning fast even for 25 rounds of robot keypad:
the (correct) answer for part 2 is now computed in 10-12 ms.