Day 22, Part 2
Generate 2000 pseudorandom numbers for each line of input,
and calculate the prices (mod 10) and price changes for each
sequence. Find which sequence of changes will yield bananas
for each secret number sequence, removing duplicates.

Merge the results for each secret number, and then find the
price change pattern that yields the largest result and return
that result.

This generates a result for part 2 in ~3.5 seconds.