Day 23, Part 1
Build a map of bidirectional edges and then loop over the
forward edges only: if any two nodes this node connects to
also connect to each other, then it's a set of three.

Filter the sets of three to find ones containing nodes that
start with 't' and return the count.