Day 24, Part 2
I completely failed to test-drive this.

The processor is an adder, and since it doesn't use
NAND gates there's probably only one kind it can
be.

Iterate through the gates looking for ones that
aren't part of a full adder design. We should find
exactly 8.

Filter out special cases with the first and final
bits.

Solution runs really quickly: 7ms for part 2.